src/app/User/user-login/user-login.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { FormsModule } from '@angular/forms';

import { UserLoginComponent } from './user-login.component';

describe('UserLoginComponent', () => {
  let component: UserLoginComponent;
  let fixture: ComponentFixture<UserLoginComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ UserLoginComponent ],
      imports: [FormsModule]
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(UserLoginComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  // it('[Email - check - Invalid] - Should check email field is not valid', () =>{
  //   let email = component.Loginform.form.controls['Useremail'];
  //   expect(email.valid).toBeFalsy();
  //   expect(component.forgotform.valid).toBeFalsy();
  //   email.setvalue('abc')
  // }


  )
});
